+++ src/Ease.cpp
@@ -4,10 +4,17 @@
 #include <sstream>
 #include <algorithm>
 #include <filesystem>
+#include <cstdlib>
 #include <vector>
 #include <cctype>
 #include <locale>
 #include <thread>
+
+#ifndef EASE_HPP
+#include "Ease.hpp"
+#endif
+
+#define NS EASE_NAMESPACE
 
 // trim from start (in place)
 void ltrim(std::string &s) {
@@ -35,10 +42,7 @@ std::string ltrim_copy(std::string s) {
 	return s;
 }
 
-#define NS Ease
-
 std::filesystem::path Default_State_File = "build/state.txt";
-std::filesystem::path Working_Directory;
 
 std::string get_compile_flag(NS::Build::Cli cli, NS::Build::Std_Ver std_ver) noexcept;
 std::string get_exe_output_flag(NS::Build::Cli cli, std::string_view str) noexcept;
@@ -77,17 +81,31 @@ NS::Flags NS::Flags::parse(int argc, char** argv) noexcept {
 		if (strcmp(it, "--debug") == 0) {
 			flags.generate_debug = true;
 		}
+		if (strcmp(it, "--install") == 0) {
+			flags.install = true;
+		}
 		if (strcmp(it, "-l") == 0 || strcmp(it, "--link") == 0) {
 			flags.link_only = true;
 		}
 		if (strcmp(it, "-h") == 0 || strcmp(it, "--help") == 0) {
 			flags.show_help = true;
 		}
+		if (strcmp(it, "--help-install") == 0) {
+			flags.show_help_install = true;
+		}
+		if (strcmp(it, "--verbose") == 0) {
+			if (i + 1 >= argc) continue;
+			++i;
+			flags.verbose_level = std::stoi(argv[i]);
+		}
 		if (strcmp(it, "--release") == 0) {
 			flags.release = true;
 		}
 		if (strcmp(it, "--scratch") == 0) {
 			flags.scratch = true;
+		}
+		if (strcmp(it, "--no-install-path") == 0) {
+			flags.no_install_path = true;
 		}
 		if (strcmp(it, "--state-file") == 0) {
 			if (i + 1 >= argc) continue;
@@ -126,6 +144,9 @@ const char* NS::Flags::help_message() noexcept {
 	"-h|--help                    Will show this message, then exit.\n"
 	"                    Exemple: ./Build.exe -h | ./Build.exe --help\n\n"
 
+	"--help-install               Print informations about the install system.\n"
+	"                    Exemple: ./Build.exe --help-install\n\n"
+	
 	"-j <n>                        Will use n threads.\n"
 	"                    Exemple: ./Build.exe -j 4\n\n"
 
@@ -133,9 +154,24 @@ const char* NS::Flags::help_message() noexcept {
 	"                             state available. Will not check for incremental compilation.\n"
 	"                    Exemple: ./Build.exe -l | ./Build.exe --link\n\n"
 
+	"--verbose <level>            Change the level of verbosity.\n"
+	"                             Level 0 (default) will show a simple progress message.\n"
+	"                             Level 1 will show with each stage the commands executed.\n"
+	"                             Level 2 will show additional informations.\n"
+	"                    Exemple: ./Build.exe --verbose 0 | ./Build.exe --verbose 1\n\n"
+
 	"--release                    Will compile the program using default option for release\n"
 	"                             (for instance, if the cli choosen is gcc, will compile using -O3)\n"
 	"                    Exemple: ./Build.exe --release\n\n"
+
+	"--install                    Will install the program to be referenced by the EASE_PATH\n"
+	"                             Environment variable, for more information you can run me with\n"
+	"                             the flag --help-env\n"
+	"                    Exemple: ./Build.exe --install\n\n"
+	
+	"--no-install-path            Will *not* search your install variable for files to include\n"
+	"                             and or directory to look for librairies.\n"
+	"                    Exemple: ./Build.exe --no-install-path\n\n"
 	
 	"--debug                      Will generate alongside the program debug symbols\n"
 	"                             It is not mutually exclusive with --release"
@@ -162,6 +198,12 @@ const char* NS::Flags::help_message() noexcept {
 	;
 }
 
+const char* NS::Flags::help_install_message() noexcept {
+	return
+	"Idk man... figure it out :/.\n"
+	;
+}
+
 // ===================== FLAGS
 
 // ===================== BUILD
@@ -183,9 +225,24 @@ void NS::Build::add_library(const std::filesystem::path& f) noexcept {
 void NS::Build::add_source(const std::filesystem::path& f) noexcept {
 	auto x = f;
 	x = x.lexically_normal();
+	if (!std::filesystem::is_regular_file(x)) {
+		printf(
+			"Can't add source %s. %s does not point to a source.\n",
+			f.filename().generic_string().c_str(),
+			f.generic_string().c_str()
+		);
+		return;
+	}
 	source_files.push_back(x);
 }
 void NS::Build::add_source_recursively(const std::filesystem::path& f) noexcept {
+	if (!std::filesystem::is_directory(f)) {
+		printf(
+			"Can't search recursively for %s, it is not a directory.\n",
+			f.generic_string().c_str()
+		);
+		return;
+	}
 	for (auto& x : std::filesystem::recursive_directory_iterator(f)) {
 		if (!std::filesystem::is_regular_file(x)) continue;
 		if (x.path().extension() != ".c" && x.path().extension() != ".cpp") continue;
@@ -209,6 +266,25 @@ void NS::Build::add_header(const std::filesystem::path& f) noexcept {
 void NS::Build::add_define(std::string str) noexcept {
 	defines.push_back(std::move(str));
 }
+
+std::filesystem::path NS::details::get_output_path(const Build& b) noexcept {
+	std::filesystem::path out = b.name;
+	if (b.flags.output) {
+		out = *b.flags.output;
+		if (std::filesystem::is_directory(*b.flags.output)) {
+			out /= b.name;
+		}
+	}
+
+	return out;
+}
+
+void NS::Build::add_default_win32() noexcept {
+	add_library("Shell32");
+	add_library("Ole32");
+}
+
+
 // ===================== BUILD
 
 // ===================== State
@@ -267,6 +343,18 @@ NS::State construct_new_state(const std::filesystem::path& p) noexcept {
 }
 
 // ===================== State
+// ===================== Commands
+
+void NS::Commands::add_command(std::string c) noexcept {
+	add_command(c, std::move(c));
+}
+
+void NS::Commands::add_command(std::string c, std::string desc) noexcept {
+	commands.emplace_back(std::move(c));
+	short_desc.emplace_back(std::move(desc));
+}
+
+// ===================== Commands
 
 
 void concat(
@@ -284,17 +372,20 @@ void append(
 	dump_to_file(to_dump, dst);
 }
 
-std::string compile_script_header_only(const NS::Build& b) noexcept {
-	std::string script;
+std::string produce_single_header(NS::Build& b) noexcept {
+	std::string single;
 
-	for (auto& x : b.header_files) script += read_file(x) + "\n";
-	if (b.invert_header_implementation_define) script += "#ifndef " + b.name + "_header_only\n";
-	else script += "#ifdef " + b.name + "_implementation\n";
+	for (auto& x : b.header_files) single += read_file(x) + "\n";
+	if (b.invert_header_implementation_define) single += "#ifndef " + b.name + "_header_only\n";
+	else single += "#ifdef " + b.name + "_implementation\n";
 	for (auto& x : b.source_files) {
-		script += read_file(x) + "\n";
+		single += read_file(x) + "\n";
 	}
-	script += "#endif\n";
-	return script;
+	single += "#endif\n";
+
+	dump_to_file(single, b.name + ".hpp");
+
+	return single;
 }
 
 NS::Commands compile_command_object(NS::State& state, const NS::Build& b) noexcept {
@@ -340,7 +431,7 @@ NS::Commands compile_command_object(NS::State& state, const NS::Build& b) noexce
 
 		command += " " + c.generic_string();
 
-		commands.commands.push_back(command);
+		commands.add_command(command, "Compile " + x.filename().generic_string());
 	}
 	return commands;
 }
@@ -364,16 +455,13 @@ NS::Commands compile_command_link_exe(const NS::Build& b) noexcept {
 	if (b.flags.generate_debug)
 		command += " " + get_debug_symbol_link_flag(b.cli);
 
-	std::string exe_name = b.name + ".exe";
-	if (b.flags.output) {
-		exe_name = b.flags.output->generic_string();
-		if (std::filesystem::is_directory(*b.flags.output)) {
-			exe_name += b.name + ".exe";
-		}
-	}
+	auto exe_path = NS::details::get_output_path(b).replace_extension("exe");
 
-	command += " " + get_exe_output_flag(b.cli, exe_name);
-	commands.commands.push_back(command);
+	commands.file_output.push_back(exe_path);
+
+	command += " " + get_exe_output_flag(b.cli, exe_path.generic_string());
+	commands.add_command(command, "Link " + b.name);
+	return commands;
 }
 
 NS::Commands compile_command_incremetal_check(const NS::Build& b) noexcept {
@@ -406,7 +494,7 @@ NS::Commands compile_command_incremetal_check(const NS::Build& b) noexcept {
 
 		command += " " + f.generic_string();
 
-		commands.commands.push_back(command);
+		commands.add_command(command, "Preprocess " + x.filename().generic_string());
 	}
 
 	return commands;
@@ -414,15 +502,54 @@ NS::Commands compile_command_incremetal_check(const NS::Build& b) noexcept {
 
 void execute(const NS::Build& build, const NS::Commands& c) noexcept {
 	std::vector<std::thread> threads;
+	std::atomic<bool> stop_flag = false;
+
 	for (size_t i = 0; i < build.flags.j; ++i) {
 		threads.push_back(std::thread([&, i] {
 			for (size_t j = i; j < c.commands.size(); j += build.flags.j) {
-				printf("%s\n", c.commands[j].c_str());
-				system(c.commands[j].c_str());
+				if (stop_flag) return;
+
+				printf(
+					"[%*d/%*d] %s\n",
+					(int)(1 + std::log10((double)c.commands.size())),
+					(int)(1 + j),
+					(int)(1 + std::log10((double)c.commands.size())),
+					(int)c.commands.size(),
+					c.short_desc[j].c_str()
+				);
+				if (build.flags.verbose_level > 0) printf("%s\n", c.commands[j].c_str());
+				auto ret = system(c.commands[j].c_str());
+			
+				if (ret != 0) stop_flag = true;
 			}
 		}));
 	}
 	for (auto& x : threads) x.join();
+
+	if (stop_flag) {
+		printf("There was an error in the build. There should be more informations above.");
+	}
+}
+
+void add_install_path(NS::Build& b) noexcept {
+	if (auto ease_path = std::getenv("EASE_PATH"); ease_path) {
+		std::vector<std::string> dirs;
+		auto start = ease_path;
+		auto prev = start;
+		auto end = ease_path + strlen(ease_path);
+		dirs.resize(std::count(start, end, ';'));
+		
+		size_t i = 0;
+		while((start = std::find(start, end, ';')) != end) {
+			dirs[i++] = std::string(prev, start - prev);
+			prev = start;
+		}
+
+		for (auto& x : dirs) if (std::filesystem::is_directory(x)) {
+			b.add_include(x);
+			b.add_library(x);
+		}
+	}
 }
 
 int main(int argc, char** argv) {
@@ -430,16 +557,28 @@ int main(int argc, char** argv) {
 	auto flags = NS::Flags::parse(argc - 1, argv + 1);
 
 	auto b = build(flags);
+	NS::Working_Directory = std::filesystem::absolute(std::filesystem::current_path());
 
 	if (flags.show_help) {
 		printf("%s", NS::Flags::help_message());
 		return 0;
 	}
+	if (flags.show_help_install) {
+		printf("%s", NS::Flags::help_install_message());
+		return 0;
+	}
+
+	if (flags.verbose_level > 1) {
+		printf("Running in %s.\n", NS::Working_Directory.generic_string().c_str());
+	}
+
 	if (flags.clean) {
 		std::filesystem::remove_all("build");
 		std::filesystem::remove_all("temp");
 		return 0;
 	}
+
+	if (!flags.no_install_path) add_install_path(b);
 
 	NS::State old_state = {};
 	NS::State new_state = {};
@@ -456,8 +595,8 @@ int main(int argc, char** argv) {
 	case NS::Build::Target::Header_Only : {
 		for (auto& x : b.pre_compile) execute(b, x);
 
-		to_dump = compile_script_header_only(b);
-		dump_to_file(to_dump, b.name + ".hpp");
+		produce_single_header(b);
+		b.to_install.push_back(NS::details::get_output_path(b).replace_extension("hpp"));
 
 		for (auto& x : b.post_compile) execute(b, x);
 		for (auto& x : b.pre_link) execute(b, x);
@@ -465,6 +604,8 @@ int main(int argc, char** argv) {
 		break;
 	}
 	case NS::Build::Target::Exe : {
+		if (b.source_files.empty()) break;
+
 		std::filesystem::create_directory("build");
 		std::filesystem::create_directory("temp");
 
@@ -484,6 +625,11 @@ int main(int argc, char** argv) {
 			execute(b, c);
 		}
 		c = compile_command_link_exe(b);
+		// in the link phase we add the executable(s) to the install path.
+		for (auto& x : c.file_output) {
+			b.to_install.push_back(x);
+		}
+
 		for (auto& x : b.pre_link) execute(b, x);
 		execute(b, c);
 		for (auto& x : b.post_link) execute(b, x);
@@ -502,6 +648,7 @@ int main(int argc, char** argv) {
 		break;
 	}
 
+	if (b.flags.install) NS::details::install_build(b);
 
 	return 0;
 }
@@ -542,10 +689,10 @@ std::string get_debug_symbol_compile_flag(NS::Build::Cli cli) noexcept {
 	}
 	return "";
 }
-std::string get_debug_symbol_compile_flag(NS::Build::Cli cli) noexcept {
+std::string get_debug_symbol_link_flag(NS::Build::Cli cli) noexcept {
 	switch (cli) {
 	case NS::Build::Cli::Gcc :
-		return "";
+		return "-g";
 		break;
 	case NS::Build::Cli::Cl :
 		return "/DEBUG";
@@ -720,10 +867,4 @@ std::string unique_name(const std::filesystem::path& path) noexcept {
 	return ret;
 }
 
-
-
 #undef NS
-
-#ifndef BS_NO_NAMESPACE
-using namespace Ease;
-#endif
